attendance_system/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy>=2.0.22
psycopg2-binary>=2.9.9
google-generativeai>=0.3.1
pydantic>=2.4.2
pydantic-settings>=2.0.3
jinja2>=3.1.2
python-dotenv>=1.0.0
passlib>=1.7.4
bcrypt==4.0.1
